Fix routes for quiz submission files
Quiz submissions are namespaced, but the routes for their files weren't.
If the Canvas installation is backed by a local file system, users will
get a Page Not Found error when attempting to view or download the file,
because the link generated by Attachment#local_storage_path

    quizzes/quiz_submissions/:quiz_id/files/:attachment_id/download[...]

does not match any routes. This fix scopes all quiz submission file
routes with the `/quizzes` prefix, so the link works once again.

Test plan:

* Create a quiz with a File Upload question
* As a student, complete the quiz by uploading a file
* As a teacher, view the submission and click the file link
* Verify the file downloads successfully with no errors